style: 파일 경로 변경 - fetcher를 utils로 변경 - 로그인 회원가입 로그아웃 method를 auth파일로 통합
